misplaced DataFrame.join test (#32375)
