+++ pandas/tests/frame/test_combine_concat.py
@@ -8,7 +8,7 @@ from pandas import DataFrame, Index, Series, Timestamp, date_range
 import pandas._testing as tm
 
 
-class TestDataFrameConcatCommon:
+class TestDataFrameConcat:
     def test_concat_multiple_frames_dtypes(self):
 
         # GH 2759
@@ -106,77 +106,6 @@ class TestDataFrameConcatCommon:
             }
         )
         tm.assert_frame_equal(results, expected)
-
-    def test_join_str_datetime(self):
-        str_dates = ["20120209", "20120222"]
-        dt_dates = [datetime(2012, 2, 9), datetime(2012, 2, 22)]
-
-        A = DataFrame(str_dates, index=range(2), columns=["aa"])
-        C = DataFrame([[1, 2], [3, 4]], index=str_dates, columns=dt_dates)
-
-        tst = A.join(C, on="aa")
-
-        assert len(tst.columns) == 3
-
-    def test_join_multiindex_leftright(self):
-        # GH 10741
-        df1 = pd.DataFrame(
-            [
-                ["a", "x", 0.471780],
-                ["a", "y", 0.774908],
-                ["a", "z", 0.563634],
-                ["b", "x", -0.353756],
-                ["b", "y", 0.368062],
-                ["b", "z", -1.721840],
-                ["c", "x", 1],
-                ["c", "y", 2],
-                ["c", "z", 3],
-            ],
-            columns=["first", "second", "value1"],
-        ).set_index(["first", "second"])
-
-        df2 = pd.DataFrame(
-            [["a", 10], ["b", 20]], columns=["first", "value2"]
-        ).set_index(["first"])
-
-        exp = pd.DataFrame(
-            [
-                [0.471780, 10],
-                [0.774908, 10],
-                [0.563634, 10],
-                [-0.353756, 20],
-                [0.368062, 20],
-                [-1.721840, 20],
-                [1.000000, np.nan],
-                [2.000000, np.nan],
-                [3.000000, np.nan],
-            ],
-            index=df1.index,
-            columns=["value1", "value2"],
-        )
-
-        # these must be the same results (but columns are flipped)
-        tm.assert_frame_equal(df1.join(df2, how="left"), exp)
-        tm.assert_frame_equal(df2.join(df1, how="right"), exp[["value2", "value1"]])
-
-        exp_idx = pd.MultiIndex.from_product(
-            [["a", "b"], ["x", "y", "z"]], names=["first", "second"]
-        )
-        exp = pd.DataFrame(
-            [
-                [0.471780, 10],
-                [0.774908, 10],
-                [0.563634, 10],
-                [-0.353756, 20],
-                [0.368062, 20],
-                [-1.721840, 20],
-            ],
-            index=exp_idx,
-            columns=["value1", "value2"],
-        )
-
-        tm.assert_frame_equal(df1.join(df2, how="right"), exp)
-        tm.assert_frame_equal(df2.join(df1, how="left"), exp[["value2", "value1"]])
 
     def test_concat_named_keys(self):
         # GH 14252
